Move files into private directory
