+++ src/browser-locale/private/getBrowserLocale.js
